src/requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==5.0.1
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
kombu==5.3.5
packaging==24.1
pika==1.3.2
pillow==10.2.0
prompt-toolkit==3.0.43
psycopg==3.1.17
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.9.0
tzdata==2023.4
vine==5.1.0
wcwidth==0.2.13
